requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic[tz]==1.14.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asgi-correlation-id==4.2.0
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
coverage[toml]==7.6.9
    # via pytest-cov
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
execnet==2.1.1
    # via pytest-xdist
fastapi[all]==0.115.6
    # via -r requirements.in
fastapi-cli[standard]==0.0.6
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.4
    # via fastapi
mako==1.3.8
    # via migrations
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.12
    # via fastapi
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
prometheus-client==0.21.1
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==7.0.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.10.3
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-extra-types==2.10.1
    # via fastapi
pydantic-settings==2.6.1
    # via
    #   -r requirements.in
    #   fastapi
pygments==2.18.0
    # via rich
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-dotenv
    #   pytest-mock
    #   pytest-xdist
pytest-asyncio==0.24.0
    # via -r requirements.in
pytest-cov==6.0.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
pytest-mock==3.14.0
    # via -r requirements.in
pytest-xdist==3.6.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   pytest-dotenv
    #   uvicorn
python-multipart==0.0.19
    # via fastapi
pyyaml==6.0.2
    # via
    #   fastapi
    #   uvicorn
requests==2.32.3
    # via -r requirements.in
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.12.0
    # via fastapi-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   migrations
starlette==0.41.3
    # via
    #   asgi-correlation-id
    #   fastapi
    #   prometheus-fastapi-instrumentator
typer==0.15.1
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   migrations
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sqlalchemy
    #   typer
ujson==5.10.0
    # via fastapi
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.32.1
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.0
    # via uvicorn
websockets==14.1
    # via uvicorn
